Move lock status icon into ListViewBlockSelectButton component
