+++ packages/block-editor/src/components/list-view/style.scss
@@ -196,18 +196,6 @@
 		}
 	}
 
-	.block-editor-list-view-block__lock-cell {
-		line-height: 0;
-		width: 24px;
-		min-width: 24px;
-		padding: 0;
-		vertical-align: middle;
-
-		.components-button:disabled {
-			opacity: 1;
-		}
-	}
-
 	.block-editor-list-view-block__mover-cell-alignment-wrapper {
 		display: flex;
 		height: 100%;
@@ -304,6 +292,15 @@
 
 	&.is-selected .block-editor-list-view-block-select-button__anchor {
 		background: rgba($black, 0.3);
+	}
+
+	.block-editor-list-view-block-select-button__lock {
+		line-height: 0;
+		width: 24px;
+		min-width: 24px;
+		margin-left: auto;
+		padding: 0;
+		vertical-align: middle;
 	}
 }
